basic project card complete
