+++ components/skillIcons.tsx
@@ -1,0 +1,28 @@
+type SkillIconMap = {
+	[key: string]: string;
+};
+
+const skillIcons: SkillIconMap = {
+	"Machine Learning": "/skills/icons8-brain-48.png",
+	Dart: "/skills/icons8-dart.svg",
+	Ruby: "/skills/icons8-ruby-programming-language.svg",
+	Latex: "/skills/icons8-latex.svg",
+	Vue: "/skills/icons8-vue-js.svg",
+	TypeScript: "/skills/icons8-typescript.svg",
+	Swift: "/skills/icons8-swift.svg",
+	Rust: "/skills/icons8-rust-programming-language.svg",
+	React: "/skills/icons8-react-native.svg",
+	Python: "/skills/icons8-python.svg",
+	Node: "/skills/icons8-nodejs.svg",
+	JavaScript: "/skills/icons8-javascript.svg",
+	Java: "/skills/icons8-java.svg",
+	HTML: "/skills/icons8-html-5.svg",
+	Golang: "/skills/icons8-golang.svg",
+	Docker: "/skills/icons8-docker.svg",
+	CSS: "/skills/icons8-css3.svg",
+	"C++": "/skills/icons8-c++.svg",
+	Csharp: "/skills/icons8-c-sharp-logo.svg",
+	C: "/skills/icons8-c-programming.svg",
+};
+
+export default skillIcons;
